requirements.txt
addict==2.4.0
aiohttp==3.8.4
aiosignal==1.3.1
airtest==1.2.2
aliyun-python-sdk-core==2.13.36
aliyun-python-sdk-kms==2.16.0
allure-pytest==2.9.43
allure-python-commons==2.9.43
altgraph==0.17
asgiref==3.3.4
async-timeout==4.0.2
asynctest==0.13.0
atomicwrites==1.4.0
attrs==21.2.0
backports.zoneinfo==0.2.1
bcrypt==3.2.0
blinker==1.4
blis==0.7.9
boto3==1.26.126
botocore==1.29.126
Brotli==1.0.9
cached-property==1.5.2
cachetools==4.2.2
catalogue==2.0.8
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==3.1.0
click==7.1.2
colorama==0.4.4
comtypes==1.1.10
confection==0.0.4
coreapi==2.3.3
coreschema==0.0.4
crcmod==1.7
cryptography==3.2.1
cssselect==1.1.0
cssutils==2.3.0
cx-Oracle==8.2.1
cycler==0.11.0
cymem==2.0.7
datasets==2.8.0
DBUtils==1.3
decorator==5.0.9
Deprecated==1.2.12
dill==0.3.6
Django==3.2.7
django-cors-headers==3.8.0
django-filter==2.4.0
django-rest-multiple-models==2.1.3
djangorestframework==3.12.4
einops==0.6.1
elasticsearch==7.12.0
en-core-web-sm==3.4.0
et-xmlfile==1.1.0
facebook-wda==1.4.3
Faker==6.6.1
falcon==3.0.1
falcon-cors==1.1.7
filelock==3.12.0
Flask==1.1.4
Flask-Cors==3.0.10
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.1.0
ftfy==6.1.1
future==0.18.2
gast==0.5.3
greenlet==1.1.2
h11==0.12.0
h2==4.1.0
hpack==4.0.0
huggingface-hub==0.14.1
hyperframe==6.0.1
idna==2.10
importlib-metadata==4.6.1
iniconfig==1.1.1
itsdangerous==1.1.0
itypes==1.2.0
jdcal==1.4.1
jieba==0.42.1
Jinja2==2.11.3
jmespath==0.10.0
joblib==1.0.1
kaitaistruct==0.9
kiwisolver==1.4.4
langcodes==3.3.0
ldap3==2.8.1
lxml==4.6.3
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.5.3
megatron-util==1.3.2
mitmproxy==5.3.0
mmdet==2.28.2
mock==5.0.2
modelscope==1.5.2
MouseInfo==0.1.3
msgpack==1.0.2
mss==4.0.3
multidict==6.0.4
multiprocess==0.70.14
murmurhash==1.0.9
mysqlclient==2.0.3
neo4j==4.2.1
nltk==3.8.1
numpy==1.21.6
opencv-contrib-python==4.5.2.54
openpyxl==3.0.5
oss2==2.17.0
packaging==21.0
pandas==1.3.5
paramiko==2.7.2
passlib==1.7.4
pathy==0.10.1
pefile==2021.9.3
Pillow==8.3.1
pluggy==0.13.1
premailer==3.10.0
preshed==3.0.8
protobuf==3.20.3
publicsuffix2==2.20191221
py==1.10.0
pyarrow==11.0.0
pyasn1==0.4.8
PyAutoGUI==0.9.53
pycocotools==2.0.6
pycparser==2.20
pycryptodome==3.10.1
pydantic==1.10.2
pydivert==2.1.0
PyGetWindow==0.0.9
pyinstaller==4.5.1
pyinstaller-hooks-contrib==2021.3
PyMsgBox==1.0.9
PyMySQL==0.10.1
PyNaCl==1.4.0
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyperclip==1.8.2
pypiwin32==223
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
PyRect==0.1.4
PyScreeze==0.1.27
pytest==6.2.4
pytest-assume==2.4.3
pytest-dependency==0.5.1
pytest-ordering==0.6
pythainlp==4.0.0
python-crfsuite==0.9.9
python-dateutil==2.8.2
PyTweening==1.0.3
pytz==2021.1
pyvi==0.1.1
pywin32==301
pywin32-ctypes==0.2.0
pywinauto==0.6.3
PyYAML==5.4.1
redis==3.5.3
regex==2022.10.31
requests==2.25.1
requests-toolbelt==0.9.1
responses==0.18.0
retry==0.9.2
rouge==1.0.1
rsa==4.9
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.7
s3transfer==0.6.0
sacremoses==0.0.53
scikit-learn==0.24.2
scipy==1.7.1
selenium==3.141.0
sentencepiece==0.1.98
seqeval==1.2.2
simplejson==3.19.1
six==1.16.0
sklearn-crfsuite==0.3.6
smart-open==6.3.0
sortedcontainers==2.2.2
spacy==3.4.4
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SQLAlchemy==1.4.27
sqlparse==0.4.2
srsly==2.4.6
subword-nmt==0.3.8
tabulate==0.9.0
termcolor==2.2.0
terminaltables==3.1.10
text-unidecode==1.3
thinc==8.1.9
threadpoolctl==2.2.0
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.1
torch==1.13.1
tornado==6.1
tqdm==4.65.0
transformers==4.28.1
typer==0.7.0
typing_extensions==4.1.1
uritemplate==3.0.1
urllib3==1.26.6
urwid==2.1.2
wasabi==0.10.1
wcwidth==0.2.6
Werkzeug==1.0.1
wincertstore==0.2
wrapt==1.12.1
wsproto==0.15.0
xlrd==1.2.0
XlsxWriter==3.0.2
xlwt==1.3.0
XMind==1.2.0
xxhash==3.2.0
yagmail==0.14.245
yapf==0.33.0
yarl==1.8.2
zhconv==1.4.3
zipp==3.5.0
zstandard==0.14.1
